Improved Suggestion Module
Fixed an error in which a Shirt could be suggested as both inner, outer, and could be paired together.
Changed the torso suggestions such that inner or outer items cannot be suggested alone if they do not match the warmth criteria. Otherwise, they can only be suggested as pairs.
Changed the prev/next/refresh methods such that they use updateClothing() to increment or decrement rather than using repetition.
Started working on unit-tests (the code is offline in notepad as it became very messy and may cause errors if pushed).